Rename AuthServiceProvider -> AuthServiceExtension which extends AbstractExtension
