--- extensions/auth/portability-auth-microsoft/src/main/resources/META-INF/services/org.dataportabilityproject.spi.gateway.auth.AuthServiceProvider
@@ -1,1 +1,0 @@
-org.dataportabilityproject.auth.microsoft.MicrosoftAuthServiceProvider
