+++ extensions/auth/portability-auth-microsoft/src/main/resources/META-INF/services/org.dataportabilityproject.spi.gateway.auth.extension.AuthServiceExtension
@@ -1,0 +1,1 @@
+org.dataportabilityproject.auth.microsoft.MicrosoftAuthServiceExtension
