+++ portability-spi-gateway/src/main/java/org/dataportabilityproject/spi/gateway/auth/AuthServiceProviderRegistry.java
@@ -1,17 +1,18 @@
 package org.dataportabilityproject.spi.gateway.auth;
 
 import java.util.Set;
+import org.dataportabilityproject.spi.gateway.auth.extension.AuthServiceExtension;
 
-/** Manages {@link AuthServiceProvider}s registered in the system. */
+/** Manages {@link AuthServiceExtension}s registered in the system. */
 public interface AuthServiceProviderRegistry {
   /**
    * Lookup the @code{AuthDataGenerator} corresponding to the serviceId, transferType and AuthMode
    * specified
    *
-   * @param serviceId the AuthServiceProvider to use
+   * @param serviceId the AuthServiceExtension to use
    * @param transferDataType The TransferdataType to authorize for
    * @param mode The authorization mode
-   * @return An AuthDataGenerator from the specified AuthServiceProvider for the type and mode
+   * @return An AuthDataGenerator from the specified AuthServiceExtension for the type and mode
    *     requested.
    */
   AuthDataGenerator getAuthDataGenerator(String serviceId, String transferDataType, AuthMode mode);
